sithara-secure/sithara-backend/requirements.txt
flask
flask-cors
flask-limiter
python-dotenv
twilio
werkzeug
